Fix bugs in test function
