+++ nomenclature/tests/test_transform_functionality.py
@@ -1,6 +1,6 @@
 import pandas as pd
 from pyam import IamDataFrame
-from transform_IamDataFrame import time_to_year_subannual
+from nomenclature import time_to_year_subannual
 
 
 TEST_DF = pd.DataFrame([
